+++ packages/solid-ui/solid-ui-components/src/ContentForm/FormCheckbox.jsx
@@ -3,8 +3,19 @@ import { Box, Label, Checkbox } from 'theme-ui'
 
 const FormCheckbox = ({ id, placeholder, ...props }) => (
   <Label htmlFor={id}>
-    <Box sx={{ position: `relative`, mr: 1 }}>
-      <Checkbox {...props} />
+    <Box
+      sx={{ position: `relative`, mr: 1, input: { zIndex: `0 !important` } }}
+    >
+      <Checkbox
+        sx={{
+          'input:checked ~ &': {
+            color: 'alpha'
+          }
+        }}
+        {...props}
+        value={undefined}
+        checked={props.value}
+      />
     </Box>
     {placeholder?.text}
   </Label>