+++ packages/solid-ui/solid-ui-theme/src/links/nav-light.js
@@ -1,8 +1,6 @@
 export default {
   color: `omegaDark`,
   textDecoration: `none`,
-  fontWeight: `bold`,
-  fontSize: 3,
   userSelect: `none`,
   cursor: `pointer`,
   whiteSpace: `pre`,