+++ packages/blocks/gatsby-blocks-core/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-blocks-core@3.8.0...@elegantstack/gatsby-blocks-core@4.0.0) (2023-03-11)
+
+
+### Features
+
+* upgrade to Gatsby v5 and update other dependencies ([aba9eb2](https://github.com/ElegantStack/gatsby-themes/commit/aba9eb243ebe39f0f02885b9693703a316c0a91f))
+
+
+### Refactor
+
+* refactor code for Gatsby v5 support ([f3852fd](https://github.com/ElegantStack/gatsby-themes/commit/f3852fd3b880cb7b77271f358b092ce730797334))
+
+
+
+
+## [3.8.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-blocks-core@3.7.0...@elegantstack/gatsby-blocks-core@3.8.0) (2022-05-18)
+
+
+### Features
+
+* Upgrade Gatsby and React and other depenedencies ([144ccc1](https://github.com/ElegantStack/gatsby-themes/commit/144ccc14a11f5f5cb5721d744922a3f164410f17))
+
+
+
+## [3.7.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-blocks-core@3.6.0...@elegantstack/gatsby-blocks-core@3.7.0) (2022-02-28)
+
+
+### Features
+
+* **solid-ui-components:** add link property to image contentType ([a57bdcb](https://github.com/ElegantStack/gatsby-themes/commit/a57bdcbcbe87a61214a47c26e5c48c420931b6fe))
+* upgrade Gatsby and other dependencies ([c42e47d](https://github.com/ElegantStack/gatsby-themes/commit/c42e47d5aea6364d7671bf5f75fbed7cee431c73))
+
+
+### Refactor
+
+* **solid-ui-blocks, gatsby-blocks-core:** hide block debug information in production mode ([782779e](https://github.com/ElegantStack/gatsby-themes/commit/782779e8ba0997d1159cc745f4f07035fde9bd6f))
+
+
+
+
+## [3.6.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-blocks-core@3.5.0...@elegantstack/gatsby-blocks-core@3.6.0) (2021-11-15)
+
+
+### Features
+
+* upgrade to Gatsby v4 ([e481e4a](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/e481e4ab705d20c9d3daf3b2048f29eef308f420))
+
+
+
+
+## [3.5.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-blocks-core@3.4.1...@elegantstack/gatsby-blocks-core@3.5.0) (2021-09-07)
+
+
+### Features
+
+* upgrade Gatsby and dependencies ([d59a731](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/d59a731107925db1a03367d9a24d7c40dae622e6))
+
+
+
+
 ### [3.4.1](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-blocks-core@3.4.0...@elegantstack/gatsby-blocks-core@3.4.1) (2021-08-03)